Split out CardListViewItem as a common widget
