+++ lib/widgets.dart
@@ -4,6 +4,7 @@ export './src/widgets/backend_version_check_navigator/backend_version_check_navi
 export './src/widgets/backend_version_incompatible_screen.dart';
 export './src/widgets/card_delete_confirm_dialog.dart';
 export './src/widgets/card_list_screen/card_list_screen.dart';
+export './src/widgets/card_list_view_item/card_list_view_item.dart';
 export './src/widgets/disappearable_builder.dart';
 export './src/widgets/fancy_popup_dialog.dart';
 export './src/widgets/fancy_popup_overlay.dart';
